backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.2.0
httpx>=0.24.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.1.0